+++ lib/karafka/web/pro/commanding/commands/trace.rb
@@ -23,6 +23,8 @@ module Karafka
                 t_d[:backtrace] = (thread.backtrace || ['<no backtrace available>']).join("\n")
               end
 
+              Dispatcher.pending(params, process_id, 'seek')
+
               Dispatcher.result(threads, process_id, 'trace')
             end
           end