+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Karafka Web Changelog
 
 ## 0.11.0 (Unreleased)
+- [Feature] Introduce a per-consumer offsets page and ability to edit offsets of running consumers (Pro).
 - [Feature] Support consumers that have mismatching schema.
 - [Feature] Provide ability to navigate to a timestamp in the Explorer (Pro).
 - [Enhancement] Provide better support for fully transactional consumers.
 - [Enhancement] Error out when `#setup` is called after `#enable!`.
+- [Change] Remove per-consumer process duplicated details from Subscriptions and Jobs tabs.
 - [Fix] The recent page breadcrumbs and offset id are not refreshed on change.
 - [Fix] Direct URL access with too big partition causes librdkafka crash.
 - [Fix] Fix incorrect breadcrumbs for pending consumer jobs.